exo/requirements.txt
# Core Agent System
autogen-agentchat>=0.2.0
langgraph>=0.0.15
openai>=1.3.0

# API Layer
fastapi>=0.104.0
uvicorn>=0.23.2
websockets>=11.0.3
pydantic>=2.4.2

# Knowledge System
chromadb>=0.4.18
neo4j>=5.14.0

# Desktop Control
pyautogui>=0.9.54

# CLI Interface
click>=8.1.7
rich>=13.6.0
textual>=0.38.1

# Testing and Development
pytest>=7.4.3
black>=23.10.1
isort>=5.12.0
